% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_final_reg_sar.R
\name{fit_quasi_newton}
\alias{fit_quasi_newton}
\title{Ré-estimação dos parâmetros com os ótimos obtidos por Newton}
\usage{
fit_quasi_newton(fit)
}
\arguments{
\item{fit}{Objeto obtido pela função fit_newton}
}
\value{
Um objeto da classe 'mle2' herdada do pacote bbmle onde
podemos obter várias estatísticas
}
\description{
Essa função serve para obter as estatísticas do modelo e
confirmar a convergência usando método de quasi-Newton implementado
na função optim. Se der erro no algoritmo BFGS, tentamos novamente
com o BFGS com restriçã (L-BFGS-B). O Objeto de saida é um modelo
completo onde se pode usar funções genéricas como coef(), summary(),
vcov(), obter intervalos de confiança e perfis de verossimilhança
}
