% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rio_proximidades}
\alias{rio_proximidades}
\title{Matriz simética de proximidades dos bairros cidade do Rio de Janeiro}
\usage{
data("rio_proximidades", package = "llsar")
}
\description{
Matriz de 0 e 1 com as ajacências dos municípios. Se um município possui adjacência com outro recebe 1 senão, recebe 0.
}
