% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rio_shape}
\alias{rio_shape}
\title{Dados de shape da cidade do Rio de Janeiro}
\usage{
data("rio_shape", package = "llsar")
}
\description{
Objeto da classe \code{SpatialPolygonsDataFrame}. Os dados de shapes de todos os bairros do Rio de Janeiro (163 bairros, em 2020) para plotar os gráficos, podem ser baixados diteramente do \href{https://www.data.rio/datasets/limite-de-bairros}{DataRio} depois importados pelo R com apoio dos pacotes \verb{rgdal e sp}.
}
