% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rio_tiroteio}
\alias{rio_tiroteio}
\title{Dados de tiroteio do Rio de Janeiro}
\usage{
data("rio_tiroteio", package = "llsar")
}
\description{
A site \href{https://fogocruzado.org.br/}{Fogo Cruzado} é uma plataforma digital colaborativa para monitoramento
civil da violência armada no Rio de Janeiro e no Recife. Usuários instalam um aplicativo no celular e
registram ocorrências de tiroteios (de maneira anônima). Os dados são disponibilizados publicamente através
de uma API, feito pela equipe do \href{https://voltdata.info/}{\code{voltdatalab}}. Depois de criar um cadastro no
\href{https://api.fogocruzado.org.br/}{site} os dados de tiroteio podem ser extraídos diretamente a API com apoio
do pacote \code{crossfire}. Para tornar os dados no nível município precisar de alguma intervenção manual.
Estes dados do pacote estão no nível bairro e já foram tratados e a consulta foi feita em novembro/2020.
}
