% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_final_reg_sar.R
\name{sar_reg_predict}
\alias{sar_reg_predict}
\title{Determina os preditos do modelo ajustado}
\usage{
sar_reg_predict(fit, idcol = "CODBAIRRO")
}
\arguments{
\item{fit}{Objeto obtido através da função fit_quasi_newton().}

\item{idcol}{Nome da variável que será a chave entre com o vetor de predições Y.}
}
\value{
data.frame com duas colunas (id e pred)
}
\description{
Como E(Y) = mu e mu é constante o vetor de valores esperados
é constante. Essa predict não é boa para o desenho do problema proposto.
}
