% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_final_reg_sar.R
\name{sar_reg_predict_oot}
\alias{sar_reg_predict_oot}
\title{Determina os preditos out-of-sampledo modelo ajustado}
\usage{
sar_reg_predict_oot(fit, idcol = "CODBAIRRO")
}
\arguments{
\item{fit}{Objeto obtido através da função fit_quasi_newton().}

\item{idcol}{Nome da variável que será a chave entre com o vetor de predições Y.}
}
\value{
data.frame com duas colunas (id e pred)
}
\description{
Essa é uma abordagem que pode ser útil quando temos apenas
uma constante como mu, mas se aplica também ao modelo de regressão
com covariáveis.
}
